% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vbHarness.R
\name{useScanVcfClo}
\alias{useScanVcfClo}
\title{illustrative closure for a variant retrieval method based on scanVcf}
\usage{
useScanVcfClo(vcffile)
}
\arguments{
\item{vcffile}{character}
}
\value{
a function with parameters \code{gr} and \code{times} for use with \code{\link{vbHarness}}
}
\description{
illustrative closure for a variant retrieval method based on scanVcf
}
