% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vbHarness.R
\name{vbHarness}
\alias{vbHarness}
\title{harness for variant access}
\usage{
vbHarness(gr, methods, times = 5)
}
\arguments{
\item{gr}{a \code{\link[GenomicRanges]{GRanges-class}} instance}

\item{times}{times a numeric passed to \code{\link[microbenchmark]{microbenchmark}} to control number of times run for averaging}
}
\description{
harness for variant access
}
